views: fix last two missing post_slug parameters
The fix here needed to use the foreign key relation that is defined for
comments <-> posts. The Comment model defines a relation here:

    post = models.ForeignKey('blog.Post', related_name='comments',
                             on_delete=models.CASCADE)

which means we can look up the post associated with each comment like so:
comment.post. This commit uses that ability to pass the post_slug parameter.